backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# PostgreSQL pgvector
pgvector==0.2.4

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# AWS S3 / MinIO
boto3==1.34.34
botocore==1.34.34

# Background tasks
celery==5.3.6
flower==2.0.1

# AI Provider SDKs
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# File parsing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
Pillow==10.2.0

# Validation & serialization
email-validator==2.1.0.post1

# CORS
flask-cors

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

